Make notifications execute the "default" action on click.
Summary:
This uses the feature introduced here [1] on KNotifications, so default
actions don't appear as a regular action button. Instead, they are
activated by clicking on the notification itself.

If the default action is not set, it uses the previous behavior of close
on click.

This changes both the notification popup and the notifications inside the
plasmoid.

[1] https://phabricator.kde.org/D4142

Test Plan: Manual testing.

Reviewers: #plasma, colomar, mart

Reviewed By: mart

Subscribers: mart, broulik, apol, mck182, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D4215