[System Tray] Part Revert "Trigger context menu on press" as this breaks xembedsniproxy
This part reverts commit b9af99140365c8c8cf4e3744af7e13b35c35fa18 which
changes plasmoids and SNI items to activate on mouse release instead of
mouse click.

This messes with xembedsniproxy which is relaying the mouse press
signals to the exembed client. Now at the time of relay, if the client
application queries the mouse state itself it's going to be wrong and
that breaks GTK.

BUG: 375930
FIXED-IN: 5.9.2

Reviewed-By: David Edmundson