Display speeds in bits per second instead of KiB/s
Summary:
This is a patch to make the Network Monitor desktop plasmoid widget display the upload and download speeds in bits per second (Mbps, Kbps or bps as the case may be) rather than the fixed "KiB/s" unit.

It appears the input data is always in KiB units so the bps part may never be called but left it in there anyway.

Reviewers: davidedmundson

Reviewed By: davidedmundson

Subscribers: broulik, sebas, Zren, davidedmundson, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D4551