Preview Centered wallpaper as PreserveAspectFit
Summary:
Currently the we create previews with QPixmapItem from KDeclrarative and
when a wallpaper is set to centered sets it to an enum value that
doesn't exist.

PreserveAspectFit displays the wallpaper in the correct proportions as
when it will be applied, but will still be big enough to make it
actually previewable.

BUG: 375675

Reviewers: #plasma, mart

Reviewed By: mart

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D4840