[lookandfeel] Add virtual keyboard to the LockScreen
Summary:
This change integrates QtVirtualKeyboard into the Lock screen UI.

A new button is added to the bottom row to toggle the appearance of the
virtual keyboard. The actual keyboard is integrated through a Loader so
that not having qtvirtualkeyboard plugin installed results in a graceful
error instead of a failing lock screen.

When the virtual keyboard is shown the remaining UI shifts upwards. This
is done through adding the stack view into a ColumnLayout. Unfortunately
this can result in an overlap with the clock. This needs some general
improvement as on e.g. my system the clock is never really visible.

To have the virtual keyboard working in the lock screen one also needs
another change to kscreenlocker_greet.

Test Plan: See screenshots

Reviewers: #plasma

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D4870