[lookandfeel] Improve positioning of Clock
Summary:
The clock so far was manually positioned in a hard coded way which could
result in the clock not being visible or only partly visible. This is the
case for example on my high dpi system with Wayland. See F2672846 for an
example.

This changes the layouting of the clock to use a ColumnLayout, with
spacer elements to ensure that the clock stays visible. It is possible
that the position of the clock now does no longer fully reflects the
ideas of the VDG.

Test Plan: Looks better on my system ;-)

Reviewers: #plasma, #vdg

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D4877