Set a longer wallpaper transition time
Summary:
Whilst using units is generally good, the wallpaper transition time is
way too short. It's a different type of animation to the others,
especially in a slideshow, where the transition is an important part of
the behaviour.

This patch loads it from a hidden configuration option with a duration
of 1000ms, if and only if animations are enabled.

Value determined through brief trial and error.

BUG: 365979

Test Plan:
Opened wallpaper. Clicked next.
Also changed static wallpaper manually, transition looked fine

Reviewers: #plasma, broulik

Reviewed By: #plasma, broulik

Subscribers: broulik, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D4370