[Notifications] Add context menu for thumbnail
Since it represents a file, right clicking it should provide similar options found in a file manager.
This allows to open the application with an application other than the default, e.g. if you
want to edit the screenshot right away. It also offers a "Copy to clipboard" action as well as
service actions, such as send as email.

Differential Revision: https://phabricator.kde.org/D4838