[sddm-theme] Add virtual keyboard support
Summary:
Just like in the lock screen this adds support for virtual keyboard.
Note: this requires adjustments in sddm, the greeter must specify the
env variable QT_IM_MODULE=qtvirtualkeyboard to have the virtual keyboard
shown.

If the virtual keyboard is not available, the button to enable it will
not be shown.

Test Plan: run in qmlscene

Reviewers: #plasma

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D5061