Keep system tray sidebar y position independent of current applet heading
Summary:
In the system tray we have a header at the top and the icons on the
left.

When there is no applet selected we want the icons to take up the full
width, because of this they have to start off under the title.

When an applet is selected we want these icons to dock to the left, but
not change their y position. The title then becomes the title of the
active applet.

If the active applet's title wraps this results in the icons moving all
over the place. See image in bug report.

This patch uses two text fields and toggles visibility so that we can
always align the icons to a consistent point

BUG: 378194

Test Plan:
Opened system tray with the arrow
Clicked on an applet and back
Opened system tray by clicking on applet

Reviewers: #plasma

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D5300