Fix losing wallpaper selection when clicking "Apply".
Summary:
Currently the selected wallpaper is deselected when the "Apply"
button is clicked, as if no wallpaper is selected. This patch
fixes that. I didn't understand what the original restoreIndex
thing was trying to do by passing the count, so I gutted it and
wrote code that makes more sense to me.

Reviewers: #plasma, mart

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D5440