Don't remove other applets in SystemTray::cleanupTask
Summary:
We can get into a state where we have applets with broken metadata.
(which is a different bug)

Someone clearly wrote this code to try and guard against accessing an
invalid metadata when checking names, but instead made it so that when
we try to remove a specific task in the system tray we accidentally also
remove up any of these applets as well, which doesn't make much sense.

CCBUG: 377050

Reviewers: #plasma, mart

Reviewed By: mart

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D5422