Avoid deep copy of image data when getting pixmaps in SNIs
Summary:
We get the data as a QByteArray we can just ref the original data,
rather than doing a deep copy of the data.

Might save a few cycles when we have an SNI animating.

Test Plan:
Ran an app that I know uses xembedsniproxy which always sends data as an
image
Still had an icon in the system tray as before

Reviewers: #plasma, markg

Reviewed By: markg

Subscribers: markg, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D5670