Fix co-installability with other DBus activated notification services
Summary:
In Plasma we install a DBus activated service to handle the case of
notifications being sent before Plasma is loaded. This file clashes with
other notification daemons and causes distro issues.

The convention is to make this file name match the DBus service name so
that we can detect conflicts, however in this case we know we have an
unavoidable conflict, but it doesn't really have an repurcisions other
than the user having an undetermined notification daemon chosen during
plasma bootup (which was the case before this patch too)

Effectively it's the same problem we have with the
org.freedesktop.FileManager1 service and Dolphin, and the same solution.

Test Plan:
Installed this
kquitapp5 plasmashell
notify-send some message. Confirmed it waited
started plasmashell. Got the message

Reviewers: #plasma, mart

Reviewed By: #plasma, mart

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D5968