Only enable MPRIS global shortcuts on demand
Summary:
The lockscreen uses the MPRIS dataengine. This leads to two applications
registring the service using the same component name which leads to the
shortcuts not working when the lock screen closes.

We want to have the applet register the shortcuts, but not every
dataengine user.

Ideally this shouldn't be in the dataengine either because DataEngines
are read only. This moves the logic to the service which is writable and
registers the shortcuts for the multiplexer there when requested from
the applet.

BUG: 380526
FIXED-IN: 5.10.1

Test Plan:
Opened media player.
Used play/pause shortcut. (the only one my keyboard has)

Reviewers: #plasma, broulik

Reviewed By: #plasma, broulik

Subscribers: broulik, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D6100