[Notifications Data Engine] Expose desktop entry and application service name icon
Now that we have "desktop-entry" available, expose this information in the data engine.
This could enable a notification implementation to show the application name or do smarter grouping.
The naming is somewhat suboptimal but "appName" and "appIcon" are already occupied and predate this.
I also didn't just want to overwrite the hints the application gives, especially not the "appIcon" which is used as notification icon.

Differential Revision: https://phabricator.kde.org/D6106