[Icon Applet] Repopulate if URL didn't change
We used to always delete the backing .desktop file if we changed a property in a Link (to
a document) hoping that setUrl would call populate. However, it won't do that if the URL
didn't change, so if you only changed the icon, the .desktop file was removed and never
re-created, breaking the applet.

BUG: 380922
FIXED-IN: 5.10.2

Differential Revision: https://phabricator.kde.org/D6148