simplify positioning code
Summary:
drop most of the positioning code, drop the animation of the
window position. position the popup in the signal handler of
visibleChanged as we are sure there the size is final and the
first expose event didn't arrive yet.
at that point we are sure the size is the final and correct one

the animation is supposed to be done by the morphingpopups effect
instead.

Test Plan:
notification positions are always correct now, both on X11 and
Wayland.
Unfortunately on x11 the notification typ doesn't seem to pass,
so they are correctly animated only in wayland

Reviewers: #plasma, davidedmundson

Reviewed By: #plasma, davidedmundson

Subscribers: apol, davidedmundson, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D6216