[lookandfeel/lockscreen] Fix state handling when clicking the keyboard's own hide button
Summary:
When clicking the hide button provided by the virtual keyboard, the
keyboard hides, but the state is not updated. This means the animation
for hide is not triggered and even more important the next call to
showHide sets a wrong state. The state assumes it is in visible, when
clicking it goes to hidden. So the virtual keyboard is not getting shown
when clicking the show virtual keyboard button.

To address this problem a change handler for keyboardActive is added
which ensures the state is correct whenever the keyboardActive value
changes.

BUG: 381833
FIXED-IN: 5.10.4

Reviewers: #plasma, sitter

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D6456