disable all test isolation for build.kde builds
not worth the effort. rely on the CI tooling to bring up everything.
hope other tests don't mess with our environment.

in KDE tests are not isolated, have no teardown, have implicit setup,
inter-depend on one another (which in turn makes them depend on execution
order). and everyone accepts that. it is depressing me to no end how shit
this is.