Show alternatives dialog immediately
Summary:
There is (a very appreciated) a description of why loading the
alternatives dialog was queued.

This causes issues for autohide as it gives us a time when the menu is
closed and we don't have anything else inhibiting autohide.

The comment doesn't seem to hold true with the latest frameworks which has
some mouse releasing code on the menu itself.

Test Plan: Clicked a lot. Clicks worked first time.

Reviewers: #plasma

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D6647