[krunner] First set panel behavior than as panel
Summary:
If a PlasmaShellSurface for an already mapped window is created and set
to panel it affects other windows by adjusting the maximized area. This
happens when showing KRunner for the second time and later. To prevent
this and make KWin's life slightly easier we first set the panel behavior
to windows go below and then mark it as panel. Thus other windows are not
affected by opening KRunner.

Reviewers: #plasma

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D6657