digital-clock: Update iso-3166 codes in timezones
Update the list of country names to the current names in iso-codes
(as output by isoquery -c).

The motivation is updating the name of Czechia as a followup to
https://git.reviewboard.kde.org/r/129644/
As suggested by Albert Astals Cid, fixing the name in i10n is not
enough:
country names are hardcoded in many places and they need to be fixed
separately.

REVIEW: 129691