[Wallpaper Configuration] Hide Get New Stuff if disabled by KIOSK
KCMs can be individually disabled by KIOSK if desired but wallpaper configuration cannot (unless you lock down
Plasma completely). Hence, at least honoring the ghns restriction here, until we have a proper solution
(that is a QtQuick equivalent of KNS3::Button).

Also, KNS3::DownloadDialog refuses to open when ghns is restricted *but* this applies only to using open()
or exec() but not a blatant show() we do from QML.

Differential Revision: https://phabricator.kde.org/D7086