Don't change fillMode of image just before deleting it
It is still set for when a source is applied for the next
transistion.

Summary:
This saves evaluating a new paintedSize and doing a bunch of evaluations
which are not used.

More importantly it works round a Qt animation crash.

fadeAnim.running is bound to otherImage.status !== loading

with the intention of delaying animations until the images are really
loaded.

By setting the fill mode and then the source, we go to loading and back,
effectively restarting the animation from within the animation.
This leads to a crash.

CCBUG: 381105

Test Plan:
End user result is the same.

Reviewers: #plasma, broulik, mart

Reviewed By: #plasma, broulik, mart

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D7248