[Notifications] Clean up States
While I'm usually a huge fan of States, this is done only half-way with States in the compact
representation (NotificationIcon) overwriting the plasmoid's state in addition to there being
a binding on it in main.qml.

It also used a "default" state (note: the default state in QML is empty string) in addition to
assigning default values on items themselves anyway.

This seems to shift evaluation of bindings into an order where we first stop the BusyIndicator
and then become passive, moving the the systray popup, working around the Qt crash we're seeing.

It's a nice clean up in any case imho.

CCBUG: 381105

Differential Revision: https://phabricator.kde.org/D7319