Don't elide UserDelegate text when there's only one item
Summary:
In the lock screen the user delegate has to elide the text when there
are multiple items next to each other, otherwise it will overlap.

When there's only one item (which is quite common) we don't need to
and doing so looks weird.

Test Plan:
Modified SDDM dummydata/userModel to have one entry with a long name
and back to multiple entries.

Reviewers: #plasma, broulik

Reviewed By: #plasma, broulik

Subscribers: broulik, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D7472