Don't set QT_SCREEN_SCALE_FACTORS on wayland
Summary:
Scale factor is sent on the output, we don't need to have the hacky env
variables.

When the kscreen KCM is used on wayland it doesn't set the variable this
is using.

Test Plan:
Configured an X session to have high DPI with this variable.
Logged into a wayland session, ran env.

(font DPI is still a problem, if it gets set on X
 it's visually doubled on my wayland session. )

Reviewers: #plasma

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D7781