Make calendar pin persistent across sessions
Summary:
A user can "pin" the calendar so that it won't get closed on focus
change, but only when the clock is clicked again.

This persists across open/close operations, but currently doesn't
persist across Plasma restarts.

This patch loads the saved config value of the last pinned state.

Test Plan:
Opened calendar, set pin, closed it, reopened and it was still pinned
(which is the current state)
killed plasma, restarted plasma, was still pinned (new)

Reviewers: #plasma

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D8252