Dict Engine: cache available dicts
This makes the config dialog more responsive when opening it multiple
times, and minimizes network traffic.

The equality check at the beginning of sourceRequestEvent was
completely wrong, if the source (DataContainer) was deleted-because-unused then
we need to proceed with the request even if it's the same as the
previous request. And if the request is ongoing then DataEngine
doesn't call us again, it connects to the existing source...