Don't set a window icon for plasmashell
Summary:
Setting a window icon is surprisingly costly and a decent chunk of
startup time. The shell doesn't need one. The only place I could
find where this icon is visible is in the KSysguard process list :).

Some people may be reminded of this rejected request:

https://git.reviewboard.kde.org/r/128484/

But that's Plasma::Dialog in plasma-framework, which (currently)
still sets an icon. This is about the shell.

Reviewers: #plasma, davidedmundson, mart

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D9196